requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
pillow